ref: general refactor
